fix: forward vLLM sampling parameters (#229)
## Description
Fix vllm backend's sampling parameters

## Motivation
After we upgraded the vLLM version, the init parameters for
`SamplingParameters` has been changed. Our previous implementation of
forwarding the parameters to vLLM backend is via filtering those
parameters not in the it, which is broken. i.e., in the current version,
if we set `max_tokens`, it will NOT be sent into the sampling parameters
(because now it is initialized from `msgspec.Struct`.

So for simplify, now we just forward all other parameters to initialize
it and return error if cannot initialize the sampling parameter.

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Documentation update

## Checklist
- [x] I have read the
[CONTRIBUTION](https://github.com/ServerlessLLM/ServerlessLLM/blob/main/CONTRIBUTING.md)
guide.
- [x] I have updated the tests (if applicable).
- [x] I have updated the documentation (if applicable).