fix: update CUDA_HOME/ROCM_HOME installation check in sllm_store (#232)
## Description

Update `setup.py` to make the CUDA_HOME/ROCM_HOME check error clearer.

Now we will import both `CUDA_HOME` and `ROCM_HOME` from
`torch.utils.cpp_extension`, and raise an error when neither of them is
detected.
 
## Motivation
As in #230, the error raised in compile `sllm_store` is unclear. It will
raise a `hip not found` error when `CUDA_HOME` is not set.

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Documentation update

## Checklist
- [x] I have read the
[CONTRIBUTION](https://github.com/ServerlessLLM/ServerlessLLM/blob/main/CONTRIBUTING.md)
guide.
- [ ] I have updated the tests (if applicable).
- [ ] I have updated the documentation (if applicable).